Updated RoleGame.java, WeaponFactory.java Added Arrow.java
In RoleGame.java I have added outlines for all of the weapons classes along with implementing the code for arrow.

In WeaponFactory.java I have changed the switch case to include all of the classes but have only implemented the arrow case so far since I have not created the Axe or Magicstaff objects yet.

The Arrow.java file now includes the code to create an arrow object just as the sword code does.

Code is functional at this point