Updated Rolegame.java and WeaponFactory.java Created Axe.java and MagicStaff.java
Axe.java is a new class that implements the weapon axe. It extends BasicWeapon and implements Weapon. The class is fully functional.

MagicStaff.java is a new class that implements the weapon magic staff. It extends BasicWeapon and implements Weapon. The class is fully functional.

Both RoleGame.java and WeaponFactory.java had code added to them that was needed for the classes implemented above.

The entire project in functional and meets the rquirements for the SoftwareEngineering homework.