Updated Axe.java file
The ARMORTHRESHOLD variable was not being used and so that was implemented and now everything is ready to be turned in.

Note: in the assignement requirements it states that the axe will ignore armor if it is less than 20. The way it is written is not inclusive of 20 and this is the way I implmented it.